HDR: Avoid u32 overflow with high pixel counts
The histogram_luminance fragment shader calculates a weighted sum of the
histogram (i.e. total brightness) using a 32-bit integer, however the
u32 can be overflowed with a high enough pixel count, resulting in
tonemap overexposure, and every pixel being pushed into bin 255 which
compounds the problem.

Since the largest histogram bin is 255, an overflow can happen if the
pixel count > 2^32/255 (~16.8MP). That corresponds to 4104x4104px, or
just 2901x2901px per eye with SBS SceneView stereo. This is easily
reached with a Varjo Aero headset, where OpenXR recommends rendering at
3604x3565 pixels per eye.

We could start using uint64 in the shader (GL_ARB_gpu_shader_int64),
however it is sufficient to just use a float for the weighted sum. The
integers are mainly only important in the histogram calculation where
pixels are being counted one by one, but at this stage we're summing
entire weighted bins worth of pixels, and with increasing weights, so
the loss of precision is likely to be minimal.

Reported-by: James Turner <james@flightgear.org>

+++ Shaders/HDR/histogram_luminance.frag
@@ -26,7 +26,7 @@ void main()
     uint light_count = histogram_get_light_count(total_pixels);
 
     int num_bins = textureSize(histogram_tex, 0).x; // [0, 255]
-    uint weighted_sum = 0u;
+    float weighted_sum = 0.0;
     uint pixel_count = 0u;
     uint valid_pixel_count = 0u;
 
@@ -47,12 +47,12 @@ void main()
         // Only consider the bins that contain pixels that are not too bright
         // or too dark for the mean.
         if (pass_dark && pass_light) {
-            weighted_sum += uint(i) * hits;
+            weighted_sum += float(i) * hits;
             valid_pixel_count += hits;
         }
     }
 
-    float mean = float(weighted_sum) / max(float(valid_pixel_count), 1.0);
+    float mean = weighted_sum / max(float(valid_pixel_count), 1.0);
 
     // Get the previous adapted luminance
     float prev_lum = max(texelFetch(prev_lum_tex, ivec2(0), 0).r, 1e-6);
